vlm-eval-pipeline/pyproject.toml
[project]
name = "vlm-eval"
version = "0.1.0"
description = "Multi-VLM evaluation pipeline for visualization literacy in immersive analytics (CS 692)"
readme = "README.md"
authors = [
    { name = "Naga Venkata Sai Chennu" },
    { name = "Hemanjali Buchireddy" }
]
requires-python = ">=3.12,<3.14"
dependencies = [
    "anthropic>=0.83.0",
    "google-genai>=1.64.0",
    "kaleido>=1.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "plotly>=6.5.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
    "tqdm>=4.67.3",
]

[project.scripts]
vlm-eval = "vlm_eval.__main__:main"

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]
